# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "The database is unavailable"
          }
        },
        "summary": "Check API and PostgreSQL readiness",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/catalog/taxonomy": {
      "get": {
        "description": "Returns the tree. By default only nodes that pass the depth gate are included, along with the ancestors needed to reach them.",
        "operationId": "CatalogController_getTaxonomy",
        "parameters": [
          {
            "name": "includeUnpublished",
            "required": false,
            "in": "query",
            "description": "Include nodes that have not yet passed the depth gate",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonomyTreeNodeDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Browse the subject taxonomy",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/nodes/{slug}": {
      "get": {
        "operationId": "CatalogController_getNode",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxonomyNodeDto"
                }
              }
            }
          },
          "404": {
            "description": "No node with that slug"
          }
        },
        "summary": "Get one taxonomy node",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/resources": {
      "get": {
        "description": "Filtering by node includes the whole subtree beneath it. Only published resources are returned.",
        "operationId": "CatalogController_listResources",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "costModel",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "free",
                "freemium",
                "paid"
              ],
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "course",
                "tutorial",
                "book",
                "reference",
                "audio",
                "video"
              ],
              "type": "string"
            }
          },
          {
            "name": "node",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceListDto"
                }
              }
            }
          }
        },
        "summary": "List published resources with facet counts",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/resources/{slug}": {
      "get": {
        "operationId": "CatalogController_getResource",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "No published resource with that slug"
          }
        },
        "summary": "Get one published resource",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/suggestions": {
      "post": {
        "description": "Deterministic ranking with per-suggestion reasons. Age-band fit gates the score; cost can only count against a resource. Gender is not accepted and cannot influence ranking.",
        "operationId": "CatalogController_suggest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerContextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionListDto"
                }
              }
            }
          },
          "400": {
            "description": "The learner context is invalid"
          },
          "404": {
            "description": "No taxonomy node with that slug"
          }
        },
        "summary": "Suggest existing catalog resources for a learner",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401": {
            "description": "A valid Auth0 access token is required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated tutor profile",
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "TutorForge API",
    "description": "Tutor-facing content platform API",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TaxonomyTreeNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scheme": {
            "type": "string",
            "enum": [
              "isced-f",
              "uk-nc",
              "sced",
              "internal"
            ]
          },
          "code": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Dot-separated node codes from the root"
          },
          "depth": {
            "type": "number"
          },
          "names": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Localised labels keyed by language tag; \"en\" is required"
          },
          "resourceCount": {
            "type": "number"
          },
          "mediaTypeCount": {
            "type": "number"
          },
          "published": {
            "type": "boolean",
            "description": "True once the node passes the depth gate"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaxonomyTreeNodeDto"
            }
          }
        },
        "required": [
          "id",
          "parentId",
          "scheme",
          "code",
          "slug",
          "path",
          "depth",
          "names",
          "resourceCount",
          "mediaTypeCount",
          "published",
          "children"
        ]
      },
      "TaxonomyNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scheme": {
            "type": "string",
            "enum": [
              "isced-f",
              "uk-nc",
              "sced",
              "internal"
            ]
          },
          "code": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Dot-separated node codes from the root"
          },
          "depth": {
            "type": "number"
          },
          "names": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Localised labels keyed by language tag; \"en\" is required"
          },
          "resourceCount": {
            "type": "number"
          },
          "mediaTypeCount": {
            "type": "number"
          },
          "published": {
            "type": "boolean",
            "description": "True once the node passes the depth gate"
          }
        },
        "required": [
          "id",
          "parentId",
          "scheme",
          "code",
          "slug",
          "path",
          "depth",
          "names",
          "resourceCount",
          "mediaTypeCount",
          "published"
        ]
      },
      "ResourceSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "course",
              "tutorial",
              "book",
              "reference",
              "audio",
              "video"
            ]
          },
          "provider": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "licence": {
            "type": "string",
            "enum": [
              "cc0",
              "public-domain",
              "cc-by",
              "cc-by-sa",
              "platform-tos",
              "proprietary"
            ]
          },
          "usageTier": {
            "type": "string",
            "enum": [
              "open",
              "embed",
              "commercial"
            ]
          },
          "costModel": {
            "type": "string",
            "enum": [
              "free",
              "freemium",
              "paid"
            ]
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          },
          "pageCount": {
            "type": "number",
            "nullable": true
          },
          "lastVerifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "mediaType",
          "provider",
          "language",
          "licence",
          "usageTier",
          "costModel",
          "durationSeconds",
          "pageCount",
          "lastVerifiedAt"
        ]
      },
      "ResourceFacetsDto": {
        "type": "object",
        "properties": {
          "mediaType": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "costModel": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          }
        },
        "required": [
          "mediaType",
          "costModel"
        ]
      },
      "ResourceListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceSummaryDto"
            }
          },
          "total": {
            "type": "number"
          },
          "facets": {
            "$ref": "#/components/schemas/ResourceFacetsDto"
          }
        },
        "required": [
          "items",
          "total",
          "facets"
        ]
      },
      "TierCapabilitiesDto": {
        "type": "object",
        "properties": {
          "deepLink": {
            "type": "boolean"
          },
          "embed": {
            "type": "boolean"
          },
          "mirror": {
            "type": "boolean"
          },
          "affiliate": {
            "type": "boolean"
          }
        },
        "required": [
          "deepLink",
          "embed",
          "mirror",
          "affiliate"
        ]
      },
      "ResourceDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "course",
              "tutorial",
              "book",
              "reference",
              "audio",
              "video"
            ]
          },
          "provider": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "licence": {
            "type": "string",
            "enum": [
              "cc0",
              "public-domain",
              "cc-by",
              "cc-by-sa",
              "platform-tos",
              "proprietary"
            ]
          },
          "usageTier": {
            "type": "string",
            "enum": [
              "open",
              "embed",
              "commercial"
            ]
          },
          "costModel": {
            "type": "string",
            "enum": [
              "free",
              "freemium",
              "paid"
            ]
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true
          },
          "pageCount": {
            "type": "number",
            "nullable": true
          },
          "lastVerifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "canonicalUrl": {
            "type": "string",
            "format": "uri"
          },
          "embedUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Only present when the usage tier permits embedding"
          },
          "outboundUrl": {
            "type": "string",
            "format": "uri",
            "description": "Where the visitor should be sent: the affiliate link where the tier permits one, otherwise the canonical URL"
          },
          "attributionText": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "capabilities": {
            "$ref": "#/components/schemas/TierCapabilitiesDto"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "description",
          "mediaType",
          "provider",
          "language",
          "licence",
          "usageTier",
          "costModel",
          "durationSeconds",
          "pageCount",
          "lastVerifiedAt",
          "canonicalUrl",
          "embedUrl",
          "outboundUrl",
          "attributionText",
          "authors",
          "isbn",
          "capabilities"
        ]
      },
      "LearnerContextDto": {
        "type": "object",
        "properties": {
          "node": {
            "type": "string",
            "description": "Taxonomy node slug to suggest within"
          },
          "ageBand": {
            "type": "string",
            "enum": [
              "5-7",
              "8-10",
              "11-13",
              "14-16",
              "17-18"
            ]
          },
          "locale": {
            "type": "string",
            "example": "en-GB",
            "description": "BCP-47 language tag"
          },
          "interests": {
            "type": "array",
            "description": "Learner interests, from a closed vocabulary",
            "items": {
              "type": "string",
              "enum": [
                "animals",
                "art",
                "baking",
                "cars",
                "coding",
                "dance",
                "dinosaurs",
                "fashion",
                "football",
                "gaming",
                "history",
                "music",
                "nature",
                "puzzles",
                "reading",
                "robotics",
                "science-fiction",
                "space",
                "sport",
                "theatre"
              ]
            }
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "default": "medium"
          },
          "learningPreference": {
            "type": "string",
            "enum": [
              "visual",
              "narrative",
              "step-by-step",
              "challenge-first"
            ],
            "default": "step-by-step"
          },
          "attentionSpan": {
            "type": "string",
            "enum": [
              "short",
              "medium",
              "long"
            ],
            "default": "medium",
            "description": "Accepted for forward compatibility; not yet used in ranking"
          },
          "allowPaid": {
            "type": "boolean",
            "default": false
          },
          "limit": {
            "type": "number",
            "default": 6,
            "minimum": 1,
            "maximum": 50
          },
          "explain": {
            "type": "string",
            "enum": [
              "top",
              "full"
            ],
            "default": "top",
            "description": "\"full\" also returns the factors that contributed nothing"
          }
        },
        "required": [
          "node",
          "ageBand",
          "locale"
        ]
      },
      "AgeGateDto": {
        "type": "object",
        "properties": {
          "factor": {
            "type": "string",
            "enum": [
              "age-band-fit"
            ]
          },
          "level": {
            "type": "string",
            "enum": [
              "exact",
              "adjacent",
              "distant",
              "none"
            ]
          },
          "multiplier": {
            "type": "number",
            "description": "Multiplies the whole weighted sum"
          }
        },
        "required": [
          "factor",
          "level",
          "multiplier"
        ]
      },
      "SuggestionReasonDto": {
        "type": "object",
        "properties": {
          "factor": {
            "type": "string",
            "enum": [
              "topic-relevance",
              "interest-overlap",
              "reading-level-fit",
              "confidence-fit",
              "media-preference",
              "quality",
              "cost"
            ]
          },
          "level": {
            "type": "string",
            "description": "Ordinal level for this factor, e.g. \"exact\""
          },
          "contribution": {
            "type": "number",
            "description": "Signed contribution to the score"
          }
        },
        "required": [
          "factor",
          "level",
          "contribution"
        ]
      },
      "SuggestionDto": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceSummaryDto"
          },
          "score": {
            "type": "number"
          },
          "gate": {
            "$ref": "#/components/schemas/AgeGateDto"
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestionReasonDto"
            }
          },
          "diversityCapped": {
            "type": "boolean",
            "description": "Pushed down by a media-type or provider cap"
          }
        },
        "required": [
          "resource",
          "score",
          "gate",
          "reasons",
          "diversityCapped"
        ]
      },
      "SuggestionListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuggestionDto"
            }
          },
          "weightsVersion": {
            "type": "string"
          },
          "candidatesConsidered": {
            "type": "number"
          },
          "emptyCause": {
            "type": "string",
            "nullable": true,
            "enum": [
              "no-candidates-in-subtree",
              "all-filtered-by-safety-vetting",
              "all-filtered-by-cost",
              "all-filtered-by-language",
              "all-filtered-by-missing-enrichment",
              "all-suppressed-by-age-gate"
            ],
            "description": "Why the list is empty; null when it is not"
          }
        },
        "required": [
          "items",
          "weightsVersion",
          "candidatesConsidered",
          "emptyCause"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt"
        ]
      }
    }
  }
}